+++ DankEngine/src/utils/memory/memoryallocator.cpp
@@ -5,13 +5,13 @@ namespace dank {
 
 	float MemoryAllocator::allocated = 0;
 
-	void* MemoryAllocator::allocate(unsigned int size) {
+	void* MemoryAllocator::allocate(size_t size) {
 		allocated += size;
 
-		byte* memory = (byte*)malloc(size + sizeof(unsigned int));
-		unsigned int* addr = (unsigned int*)memory;
+		byte* memory = (byte*)malloc(size + sizeof(size_t));
+		size_t* addr = (size_t*)memory;
 		*addr = size;
-		memory += sizeof(unsigned int);
+		memory += sizeof(size_t);
 		return (void*)memory;
 	}
 
@@ -20,8 +20,8 @@ namespace dank {
 			return; //hackerman
 
 		byte* addr = (byte*)memory;
-		addr -= sizeof(unsigned int);
-		unsigned int size = *(unsigned int*)addr;
+		addr -= sizeof(size_t);
+		size_t size = *(size_t*)addr;
 		free(addr);
 
 		allocated -= size;
